requirements.txt
alembic==1.4.2
click==7.1.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
Kivy-Garden==0.1.4
kivy-garden.mapview==1.0.5
mapview==1.0.5

pytest~=5.4.1
kivymd~=0.104.1
Kivy~=1.11.1
plyer~=1.4.3